ODROID: Initial release odroid-player
Although this project was based on yocto/linux/multimedia/gst_agmplayer
(branch: yocto-kirkstone-bsp-202409), split it to seperate repository
because overall project structure and lots of code have changed.

This commit indicates the first version of odroid-player is released.
Currently, this player supports for ODROID-C5, more ODROID boards
will be supported in the future.

This project contains one media server (odroid-player)
and four clients for now:
- odroid-player-client      : CLI interface client
- odroid-player-client-web  : Web interface client
- odroid-player-client-ir   : IR interface client
- odroid-player-client-oled : Display media information on I2C screen

This player has been tested about one week for playing videos
continuously on the ODROID-C5, and it seems no problems on playback.

Signed-off-by: Phillip Choi <phillip.choi@hardkernel.com>
Change-Id: I9acdb679a933a73bd902fe5cd1e43a93ca553375